test: improve liveApps part
