+++ apps/ledger-live-mobile/e2e/page/index.ts
@@ -6,7 +6,7 @@ import CommonPage from "./common.page";
 import CryptoDrawer from "./liveApps/cryptoDrawer";
 import CustomLockscreenPage from "./stax/customLockscreen.page";
 import DiscoverPage from "./discover/discover.page";
-import LiveAppWebview from "./liveApps/liveAppWebview";
+import DummyWalletApp from "./liveApps/dummyWalletApp.webView";
 import ManagerPage from "./manager/manager.page";
 import MarketPage from "./market/market.page";
 import NftGalleryPage from "./wallet/nftGallery.page";
@@ -37,7 +37,7 @@ export class Application {
   public cryptoDrawer = new CryptoDrawer();
   public customLockscreen = new CustomLockscreenPage();
   public discover = new DiscoverPage();
-  public liveAppWebview = new LiveAppWebview();
+  public dummyWalletApp = new DummyWalletApp();
   public manager = new ManagerPage();
   public market = new MarketPage();
   public nftGallery = new NftGalleryPage();
